rename(eligibility): 잉벤트 검증 객체 클래스로 변경
