+++ __tests__/EligibilityChecker/isEligibileForFreeGiftTest.js
