+++ __tests__/EligibilityChecker/isEligibleForEventBadgeTest.js
