+++ src/service/EligibilityChecker.js
